add assignments to top level
