+++ src/reducers/clients.js
@@ -1,20 +1,9 @@
 import Immutable from 'immutable';
 
-import {Assignment} from '../records';
-
 const defaultState = new Immutable.Map({
   firebase: new Immutable.Map({exportingSnapshot: false}),
   projectExports: new Immutable.Map(),
-  courses: new Immutable.List(),
-  assignments: new Immutable.Map(),
 });
-
-function addAssignment(state, assignment) {
-  return state.setIn(
-    ['assignments', assignment.assignmentKey],
-    Assignment.fromJS(assignment),
-  );
-}
 
 export default function clients(stateIn, action) {
   let state = stateIn;
@@ -49,33 +38,6 @@ export default function clients(stateIn, action) {
         ['projectExports', action.payload.exportType],
         new Immutable.Map({status: 'error'}),
       );
-
-    case 'UPDATE_COURSES':
-      return state.setIn(
-        ['courses'],
-        new Immutable.List(action.payload.courses),
-      );
-
-    case 'ASSIGNMENT_IMPORTED':
-      return state.setIn(
-        ['assignments', action.payload.assignment.assignmentKey],
-        new Assignment(action.payload.assignment),
-      );
-
-    case 'ASSIGNMENT_CREATED':
-      return state.setIn(
-        ['assignments', action.payload.assignmentKey],
-        new Assignment({
-          courseId: action.payload.assignment.courseId,
-          id: action.payload.assignment.id,
-          snapshotKey: action.payload.snapshotKey,
-          alternateLink: action.payload.assignment.alternateLink,
-          assignerId: action.payload.assignerId,
-        }),
-      );
-    case 'ASSIGNMENTS_LOADED':
-      return action.payload.reduce(addAssignment, state);
   }
-
   return state;
 }
